requirements.txt
Pillow==7.0.0
scipy==1.4.1
numpy==1.18.5
cmake==3.12.0
dlib==19.18.0
